Day01 Profile App
#FirstApp
This is the first app I have built through Flutter. So What is this app about?
This is the simple profile app which contain photo and details of that object. Profile App contains stateless widget so once your set profile of somebody, it can't be changed as subject inside body in stateless contain will not change.
I have built simple profile of Actor Ram Charan and the details given are all made up.

This is Aayush Basnet starting his Flutter Journey.